In TestGrainInterfaces, remove dependency on Orleans.Core
