+++ test/TestGrainInterfaces/CustomPlacement.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using Orleans;
 using Orleans.Placement;
 using Orleans.Runtime;
-using Orleans.Runtime.Placement;
 
 namespace UnitTests.GrainInterfaces
 {
@@ -60,29 +58,6 @@ namespace UnitTests.GrainInterfaces
             base(scenario == CustomPlacementScenario.FixedSilo ? TestCustomPlacementStrategy.FixedSilo : TestCustomPlacementStrategy.ExcludeOne)
         {
             Scenario = scenario;
-        }
-    }
-
-    public class TestPlacementStrategyFixedSiloDirector : IPlacementDirector<TestCustomPlacementStrategy>
-    {
-
-        public Task<SiloAddress> OnAddActivation(PlacementStrategy strategy, PlacementTarget target, IPlacementContext context)
-        {
-            var silos = context.GetCompatibleSilos(target).OrderBy(s => s).ToArray();
-            var oddTick = DateTime.UtcNow.Ticks % 2 == 1;
-
-            switch (((TestCustomPlacementStrategy)strategy).Scenario)
-            {
-                case CustomPlacementScenario.FixedSilo:
-                    return Task.FromResult(silos[silos.Length - 2]); // second from last silos.
-
-                case CustomPlacementScenario.ExcludeOne:
-                    return Task.FromResult(oddTick ? silos[0] : silos[silos.Length - 1]); // randomly return first or last silos
-
-                default:
-                    throw new InvalidOperationException(); // should never get here, only to make compiler happy
-            }
-
         }
     }
 }
