+++ test/TestGrainInterfaces/IValueTypeTestGrain.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Orleans;
 using Orleans.Concurrency;
-using Orleans.Runtime.Configuration;
 
 namespace UnitTests.GrainInterfaces
 {
