+++ test/Versions/TestVersionGrains2/TestVersionGrains2.csproj
@@ -19,5 +19,6 @@
     <ProjectReference Include="..\..\..\src\Orleans.CodeGeneration\Orleans.CodeGeneration.csproj" />
     <ProjectReference Include="..\..\..\src\Orleans.TestingHost\Orleans.TestingHost.csproj" />
     <ProjectReference Include="..\..\TestGrainInterfaces\TestGrainInterfaces.csproj" />
+    <ProjectReference Include="..\..\TestGrains\TestGrains.csproj" />
   </ItemGroup>
 </Project>
